+++ grails-app/conf/application.yml
@@ -2,7 +2,7 @@
 grails:
     profile: web
     codegen:
-        defaultPackage: com.l.w.mini.asaas
+        defaultPackage: com.miniasaaslw
     gorm:
         reactor:
             # Whether to translate GORM events into Reactor events
