--- grails-app/domain/com/l/w/mini/asaas/Customer.groovy
@@ -1,10 +1,0 @@
-package com.l.w.mini.asaas
-
-import utils.entity.BasePerson
-
-class Customer extends BasePerson {
-
-    static constraints = {
-
-    }
-}
