+++ grails-app/domain/com/miniasaaaslw/domain/customer/Customer.groovy
@@ -1,0 +1,10 @@
+package com.miniasaaaslw.domain.customer
+
+import com.miniasaaslw.utils.entity.BasePerson
+
+class Customer extends BasePerson {
+
+    static constraints = {
+
+    }
+}
