+++ grails-app/init/com/miniasaaaslw/BootStrap.groovy
@@ -1,4 +1,4 @@
-package com.l.w.mini.asaas
+package com.miniasaaaslw
 
 class BootStrap {
