+++ src/main/groovy/com/miniasaaslw/utils/entity/enums/PersonType.groovy
@@ -1,4 +1,4 @@
-package utils.entity
+package com.miniasaaslw.utils.entity.enums
 
 enum PersonType {
     FISICA,
